@@ -3,7 +3,7 @@
     "description": "Przelewy24 PHP library",
     "keywords": [
         "przelewy24",
-        "payment"
+        "payments"
     ],
     "type": "library",
     "license": "MIT",
@@ -15,11 +15,12 @@
         }
     ],
     "require": {
-        "php": ">=7.1.3",
-        "guzzlehttp/guzzle": "^6.4|^7.0"
+        "php": "^8.1",
+        "guzzlehttp/guzzle": "^7.6"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0"
+        "friendsofphp/php-cs-fixer": "^3.10",
+        "phpunit/phpunit": "^10.0"
     },
     "autoload": {
         "psr-4": {
